test_edb sorting
