+++ tests/system/test_edb_components.py
@@ -700,3 +700,13 @@ class TestClass(BaseTestClass):
         assert edbapp.export_gds_comp_xml(["U1", "U2", "C2", "R1"], control_path=xml_output)
         assert os.path.isfile(xml_output)
         edbapp.close(terminate_rpc_session=False)
+
+    def test_create_rlc_component(self):
+        """Create RLC components from pins."""
+        edb = self.edb_examples.get_si_verse()
+        pins = edb.components.get_pin_from_component("U1", "1V0")
+        ref_pins = edb.components.get_pin_from_component("U1", "GND")
+        assert edb.components.create([pins[0], ref_pins[0]], "test_0rlc", r_value=1.67, l_value=1e-13, c_value=1e-11)
+        assert edb.components.create([pins[0], ref_pins[0]], "test_1rlc", r_value=None, l_value=1e-13, c_value=1e-11)
+        assert edb.components.create([pins[0], ref_pins[0]], "test_2rlc", r_value=None, c_value=1e-13)
+        edb.close(terminate_rpc_session=False)
