--- tests/system/test_edb_differential_pairs.py
@@ -1,43 +1,0 @@
-# Copyright (C) 2023 - 2026 ANSYS, Inc. and/or its affiliates.
-# SPDX-License-Identifier: MIT
-#
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Tests related to Edb differential pairs"""
-
-import pytest
-
-from tests.system.base_test_class import BaseTestClass
-
-pytestmark = [pytest.mark.system, pytest.mark.legacy]
-
-
-@pytest.mark.usefixtures("close_rpc_session")
-class TestClass(BaseTestClass):
-    def test_differential_pairs_queries(self):
-        """Evaluate differential pairs queries"""
-        # Done
-        edbapp = self.edb_examples.get_si_verse()
-        edbapp.differential_pairs.auto_identify()
-        diff_pair = edbapp.differential_pairs.create("new_pair1", "PCIe_Gen4_RX1_P", "PCIe_Gen4_RX1_N")
-        assert diff_pair.positive_net.name == "PCIe_Gen4_RX1_P"
-        assert diff_pair.negative_net.name == "PCIe_Gen4_RX1_N"
-        assert edbapp.differential_pairs.items["new_pair1"]
-        edbapp.close(terminate_rpc_session=False)
