+++ tests/system/test_edb_modeler.py
@@ -32,6 +32,13 @@ from pyedb.generic.settings import settings
 from tests.conftest import config, local_path, test_subfolder
 from tests.system.base_test_class import BaseTestClass
 
+if config["use_grpc"]:
+    from pyedb.grpc.database.geometry.point_data import PointData
+    from pyedb.grpc.database.geometry.polygon_data import PolygonData
+else:
+    from pyedb.dotnet.database.geometry.point_data import PointData
+    from pyedb.dotnet.database.geometry.polygon_data import PolygonData
+
 
 @pytest.mark.usefixtures("close_rpc_session")
 class TestClass(BaseTestClass):
@@ -784,4 +791,50 @@ class TestClass(BaseTestClass):
             open_traces_net_filter=["SFPA_VCCR", "SFPA_VCCT"],
         )
         assert len(edbapp.layout.find_primitive(layer_name="SM")) == 4
+        edbapp.close(terminate_rpc_session=False)
+
+
+# Geometry primitives (moved from test_edb_database_geometry.py)
+@pytest.mark.usefixtures("close_rpc_session")
+class TestPointData(BaseTestClass):
+    def test_create(self):
+        edbapp = self.edb_examples.create_empty_edb()
+        edbapp["X"] = 1
+        pdata = PointData.create(edbapp, "X", 2)
+        assert str(pdata.x) == "X"
+        assert pdata.x == 1
+        assert pdata.y == 2
+
+        pdata2 = PointData.create_arc_point(edbapp, "X")
+        assert str(pdata2.arc_height) == "X"
+        assert pdata2.arc_height == 1
+        assert pdata2.is_arc
+
+        edbapp.close(terminate_rpc_session=False)
+
+    def test_operations(self):
+        edbapp = self.edb_examples.create_empty_edb()
+        edbapp["X"] = 1
+        edbapp["Y"] = 2
+        edbapp["angle"] = "90deg"
+        pdata = PointData.create(edbapp, "X", "Y")
+        pdata2 = pdata.rotate("angle", [1, 1])
+        assert pdata2.x == pytest.approx(0)
+        assert pdata2.y == pytest.approx(1)
+
+        edbapp.close(terminate_rpc_session=False)
+
+
+@pytest.mark.usefixtures("close_rpc_session")
+class TestPolygonData(BaseTestClass):
+    def test_create(self):
+        edbapp = self.edb_examples.create_empty_edb()
+        edbapp.stackup.create_symmetric_stackup(2)
+        edbapp["X"] = 1
+        pdata1 = PointData.create(edbapp, "X", 2)
+        pdata2 = PointData.create(edbapp, "X", 3)
+        pdata3 = PointData.create(edbapp, 2, 3)
+
+        poly_data = PolygonData.create(edbapp, points=[pdata1, pdata2, pdata3])
+        assert poly_data
         edbapp.close(terminate_rpc_session=False)
