+++ tests/system/test_edb_simulation_setups.py
@@ -20,158 +20,21 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-"""Tests related to Edb"""
+"""Tests related to Edb simulation setups (HFSS, SIWave, RaptorX, Q3D)."""
 
-import os
-from pathlib import Path
-from unittest.mock import MagicMock, patch
-
-import ansys.edb.core
 import pytest
 
-from pyedb.edb_logger import EdbLogger
 from pyedb.generic.general_methods import is_linux
 from tests.conftest import config
 from tests.system.base_test_class import BaseTestClass
 
 pytestmark = [pytest.mark.system, pytest.mark.grpc]
 
-ON_CI = os.environ.get("CI", "false").lower() == "true"
-
 
 @pytest.mark.usefixtures("close_rpc_session")
 class TestClass(BaseTestClass):
-    def test_hfss_create_coax_port_on_component_from_hfss(self):
-        """Create a coaxial port on a component from its pin."""
-        edbapp = self.edb_examples.get_si_verse()
-        assert edbapp.excitation_manager.create_coax_port_on_component("U1", "DDR4_DQS0_P")
-        assert edbapp.excitation_manager.create_coax_port_on_component("U1", ["DDR4_DQS0_P", "DDR4_DQS0_N"], True)
-        edbapp.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_layout_bounding_box(self):
-        """Evaluate layout bounding box"""
-        edbapp = self.edb_examples.get_si_verse()
-        assert len(edbapp.get_bounding_box()) == 2
-        bbox = [[round(i, 6) for i in j] for j in edbapp.get_bounding_box()]
-        assert bbox == [[-0.014260, -0.004550], [0.150105, 0.080000]]
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_create_circuit_port_on_net(self):
-        """Create a circuit port on a net."""
-        edbapp = self.edb_examples.get_si_verse()
-        initial_len = len(edbapp.padstacks.pingroups)
-        assert edbapp.excitation_manager.create_circuit_port_on_net("U1", "1V0", "U1", "GND", 50, "test") == "test"
-        p2 = edbapp.excitation_manager.create_circuit_port_on_net("U1", "PLL_1V8", "U1", "GND", 50, "test")
-        assert p2 != "test" and "test" in p2
-        pins = edbapp.components.get_pin_from_component("U1")
-        p3 = edbapp.excitation_manager.create_circuit_port_on_pin(pins[200], pins[0], 45)
-        assert p3 != ""
-        p4 = edbapp.excitation_manager.create_circuit_port_on_net("U1", "USB3_D_P")
-        assert len(edbapp.padstacks.pingroups) == initial_len + 6
-        assert "GND" in p4 and "USB3_D_P" in p4
-        assert "test" in edbapp.terminals
-        assert edbapp.siwave.create_pin_group_on_net("U1", "1V0", "PG_V1P0_S0")
-        assert edbapp.siwave.create_pin_group_on_net("U1", "GND", "U1_GND")
-        assert edbapp.excitation_manager.create_circuit_port_on_pin_group(
-            "PG_V1P0_S0", "U1_GND", impedance=50, name="test_port"
-        )
-        edbapp.ports["test_port"].name = "test_rename"
-        assert any(port for port in list(edbapp.ports) if port == "test_rename")
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_create_voltage_source(self):
-        """Create a voltage source."""
-        edbapp = self.edb_examples.get_si_verse()
-        assert "Vsource_" in edbapp.excitation_manager.create_voltage_source_on_net(
-            "U1", "USB3_D_P", "U1", "GND", 3.3, 0
-        )
-        assert len(edbapp.terminals) == 2
-        assert list(edbapp.terminals.values())[0].magnitude == 3.3
-
-        pins = edbapp.components.get_pin_from_component("U1")
-        assert "VSource_" in edbapp.excitation_manager.create_voltage_source_on_pin(
-            pins[300], pins[10], voltage_value=3.3, phase_value=1
-        )
-        assert len(edbapp.terminals) == 4
-        assert list(edbapp.terminals.values())[2].phase == 1.0
-        assert list(edbapp.terminals.values())[2].magnitude == 3.3
-
-        u6 = edbapp.components["U6"]
-        term1 = u6.pins["F2"].get_terminal(create_new_terminal=True)
-        term2 = u6.pins["F1"].get_terminal(create_new_terminal=True)
-        voltage_source = edbapp.create_voltage_source(terminal=term1, ref_terminal=term2)
-        assert not voltage_source.is_null
-        # testing source to ground assignment
-        setup = edbapp.siwave.add_siwave_dc_analysis(name="Test_dc")
-        setup.settings.add_source_terminal_to_ground("Vsource_U1_USB3_D_P_U1_GND", 1)
-        if edbapp.grpc:
-            assert "Vsource_U1_USB3_D_P_U1_GND" in setup.settings.source_terms_to_ground
-        else:
-            assert "Vsource_U1_USB3_D_P_U1_GND" in setup.settings.dc_ir.source_terms_to_ground
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_create_current_source(self):
-        """Create a current source."""
-        edbapp = self.edb_examples.get_si_verse()
-        assert edbapp.excitation_manager.create_current_source_on_net("U1", "USB3_D_N", "U1", "GND", 0.1, 0)
-        pins = edbapp.components.get_pin_from_component("U1")
-        assert "I22" == edbapp.excitation_manager.create_current_source_on_pin(pins[301], pins[10], 0.1, 0, "I22")
-
-        assert edbapp.siwave.create_pin_group_on_net(reference_designator="U1", net_name="GND", group_name="gnd")
-        edbapp.siwave.create_pin_group(reference_designator="U1", pin_numbers=["A27", "A28"], group_name="vrm_pos")
-        edbapp.excitation_manager.create_current_source_on_pin_group(
-            pos_pin_group_name="vrm_pos", neg_pin_group_name="gnd", name="vrm_current_source"
-        )
-
-        edbapp.siwave.create_pin_group(reference_designator="U1", pin_numbers=["R23", "P23"], group_name="sink_pos")
-        edbapp.siwave.create_pin_group_on_net(reference_designator="U1", net_name="GND", group_name="gnd2")
-
-        # move to source_excitations when dotnet is removed
-        assert edbapp.excitation_manager.create_voltage_source_on_pin_group(
-            "sink_pos", "gnd2", name="vrm_voltage_source"
-        )
-        edbapp.siwave.create_pin_group(reference_designator="U1", pin_numbers=["A27", "A28"], group_name="vp_pos")
-        assert edbapp.siwave.create_pin_group_on_net(reference_designator="U1", net_name="GND", group_name="vp_neg")
-        assert edbapp.siwave.pin_groups["vp_pos"]
-        assert edbapp.siwave.pin_groups["vp_pos"].pins
-        assert edbapp.siwave.create_voltage_probe_on_pin_group("vprobe", "vp_pos", "vp_neg")
-        assert edbapp.terminals["vprobe"]
-        edbapp.siwave.place_voltage_probe(
-            "vprobe_2", "1V0", ["112mm", "24mm"], "1_Top", "GND", ["112mm", "27mm"], "Inner1(GND1)"
-        )
-        vprobe_2 = edbapp.terminals["vprobe_2"]
-        ref_term = vprobe_2.reference_terminal
-        assert isinstance(ref_term.location, list) or isinstance(ref_term.location, tuple)
-        # ref_term.location = [0, 0] # position setter is crashing check pyedb-core bug #431
-        assert ref_term.layer
-        ref_term.layer.name = "Inner1(GND1"
-        ref_term.layer.name = "test"
-        assert "test" in edbapp.stackup.layers
-        u6 = edbapp.components["U6"]
-        assert edbapp.create_current_source(
-            u6.pins["H8"].get_terminal(create_new_terminal=True), u6.pins["G9"].get_terminal(create_new_terminal=True)
-        )
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_create_dc_terminal(self):
-        """Create a DC terminal."""
-        edbapp = self.edb_examples.get_si_verse()
-        assert edbapp.siwave.create_dc_terminal("U1", "DDR4_DQ40", "dc_terminal1") == "dc_terminal1"
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_create_resistors_on_pin(self):
-        """Create a resistor on pin."""
-        edbapp = self.edb_examples.get_si_verse()
-        pins = edbapp.components.get_pin_from_component("U1")
-        assert "RST4000" == edbapp.excitation_manager.create_resistor_on_pin(pins[302], pins[10], 40, "RST4000")
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_siwave_add_syz_analsyis(self):
-        """Add a sywave AC analysis."""
+    def test_siwave_add_syz_analysis(self):
+        """Add a SIWave AC analysis."""
         edbapp = self.edb_examples.get_si_verse()
         syz_setup = edbapp.siwave.add_siwave_syz_analysis(start_freq="1GHz", stop_freq="10GHz", step_freq="10MHz")
         syz_setup.use_custom_settings = False
@@ -190,7 +53,7 @@ class TestClass(BaseTestClass):
         edbapp.close(terminate_rpc_session=False)
 
     def test_siwave_add_dc_analysis(self):
-        """Add a sywave DC analysis."""
+        """Add a SIWave DC analysis."""
         edbapp = self.edb_examples.get_si_verse()
         assert edbapp.siwave.add_siwave_dc_analysis(name="Test_dc")
         edbapp.close(terminate_rpc_session=False)
@@ -208,220 +71,8 @@ class TestClass(BaseTestClass):
         assert len(mesh_ops) > 0
         edbapp.close(terminate_rpc_session=False)
 
-    def test_add_variables(self):
-        """Add design and project variables."""
-        edbapp = self.edb_examples.get_si_verse()
-        edbapp.add_design_variable("my_variable", "1mm")
-        assert "my_variable" in edbapp.get_all_variable_names()
-        assert edbapp.modeler.parametrize_trace_width("DDR4_DQ25")
-        assert edbapp.modeler.parametrize_trace_width("DDR4_A2")
-        if edbapp.grpc:
-            edbapp.add_design_variable("my_parameter", "2mm", "test description")
-        else:
-            edbapp.add_design_variable("my_parameter", "2mm", True)
-        assert "my_parameter" in edbapp.get_all_variable_names()
-        variable_value = edbapp.get_variable("my_parameter")
-        assert variable_value.value == 2e-3
-        if edbapp.grpc:
-            assert not edbapp.add_design_variable("my_parameter", "2mm", "test description")
-        else:
-            # grpc and DotNet variable implementation server are too different.
-            assert not edbapp.add_design_variable("my_parameter", "2mm", True)[0]
-        edbapp.add_project_variable("$my_project_variable", "3mm")
-        assert edbapp.get_variable("$my_project_variable").value == 3e-3
-        if edbapp.grpc:
-            assert not edbapp.add_project_variable("$my_project_variable", "3mm")
-        else:
-            # grpc and DotNet variable implementation server are too different.
-            assert not edbapp.add_project_variable("$my_project_variable", "3mm")[0]
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_save_edb_as(self):
-        """Save edb as some file."""
-        edbapp = self.edb_examples.create_empty_edb()
-        target_path = Path(self.edb_examples.test_folder) / "si_verse_new.aedb"
-        assert edbapp.save_as(target_path)
-        assert (target_path / "edb.def").exists()
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_export_3d(self):
-        """Export EDB to HFSS."""
-        mock_process = MagicMock()
-        edb = self.edb_examples.create_empty_edb()
-        options_config = {"UNITE_NETS": 1, "LAUNCH_Q3D": 0}
-        out = edb.write_export3d_option_config_file(self.edb_examples.test_folder, options_config)
-        assert Path(out).exists()
-        with patch("subprocess.run", return_value=mock_process) as mock_run:
-            executable = "siwave" if is_linux else "siwave.exe"
-
-            edb.export_hfss(None)
-            popen_args, _ = mock_run.call_args
-            input_cmd = popen_args[0]
-
-            input_cmd_ = [
-                str(Path(edb.ansys_em_path) / executable),
-                "-RunScriptAndExit",
-                str(Path(edb.edbpath).parent / "export_cad.py"),
-            ]
-            assert input_cmd == input_cmd_  # if is_linux else " ".join(input_cmd_)
-
-            edb.export_q3d(None)
-            popen_args, _ = mock_run.call_args
-            input_cmd = popen_args[0]
-
-            input_cmd_ = [
-                str(Path(edb.ansys_em_path) / executable),
-                "-RunScriptAndExit",
-                str(Path(edb.edbpath).parent / "export_cad.py"),
-            ]
-            assert input_cmd == input_cmd_  # if is_linux else " ".join(input_cmd_)
-
-            edb.export_maxwell(None)
-            popen_args, _ = mock_run.call_args
-            input_cmd = popen_args[0]
-
-            input_cmd_ = [
-                str(Path(edb.ansys_em_path) / executable),
-                "-RunScriptAndExit",
-                str(Path(edb.edbpath).parent / "export_cad.py"),
-            ]
-            assert input_cmd == input_cmd_  # if is_linux else " ".join(input_cmd_)
-
-        edb.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_create_edge_port_on_polygon(self):
-        """Create lumped and vertical port."""
-        target_path = self.edb_examples.copy_test_files_into_local_folder("TEDB/edge_ports.aedb")[0]
-        edb = self.edb_examples.load_edb(target_path)
-        poly_list = [poly for poly in edb.layout.primitives if poly.primitive_type == "polygon"]
-        port_poly = [poly for poly in poly_list if poly.id == 17][0]
-        ref_poly = [poly for poly in poly_list if poly.id == 19][0]
-        port_location = [-65e-3, -13e-3]
-        ref_location = [-63e-3, -13e-3]
-        if edb.grpc:
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly,
-                reference_polygon=ref_poly,
-                terminal_point=port_location,
-                reference_point=ref_location,
-            )
-        else:
-            # method already deprecated in grpc.
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly,
-                reference_polygon=ref_poly,
-                terminal_point=port_location,
-                reference_point=ref_location,
-            )
-        port_poly = [poly for poly in poly_list if poly.id == 23][0]
-        ref_poly = [poly for poly in poly_list if poly.id == 22][0]
-        port_location = [-65e-3, -10e-3]
-        ref_location = [-65e-3, -10e-3]
-        if edb.grpc:
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly,
-                reference_polygon=ref_poly,
-                terminal_point=port_location,
-                reference_point=ref_location,
-            )
-        else:
-            # method already deprecated in grpc.
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly,
-                reference_polygon=ref_poly,
-                terminal_point=port_location,
-                reference_point=ref_location,
-            )
-        port_poly = [poly for poly in poly_list if poly.id == 25][0]
-        port_location = [-65e-3, -7e-3]
-        if edb.grpc:
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly, terminal_point=port_location, reference_layer="gnd"
-            )
-        else:
-            # method already deprecated in grpc.
-            assert edb.excitation_manager.create_edge_port_on_polygon(
-                polygon=port_poly, terminal_point=port_location, reference_layer="gnd"
-            )
-        sig = edb.modeler.create_trace([[0, 0], ["9mm", 0]], "sig2", "1mm", "SIG", "Flat", "Flat")
-        assert sig.create_edge_port("pcb_port_1", "end", "Wave", None, 8, 8)
-        assert sig.create_edge_port("pcb_port_2", "start", "Gap")
-        gap_port = edb.ports["pcb_port_2"]
-        if edb.grpc:
-            assert edb.ports["pcb_port_1"].is_wave_port
-            assert not edb.ports["pcb_port_2"].is_wave_port
-            assert gap_port.component.is_null
-            assert not gap_port.is_circuit_port
-        else:
-            assert not gap_port.component
-        assert gap_port.source_amplitude == 0.0
-        assert gap_port.source_phase == 0.0
-        assert gap_port.impedance
-        # temp
-        from ansys.edb.core.database import ProductIdType
-
-        assert not gap_port.deembed
-        gap_port.name = "gap_port"
-        assert gap_port.name == "gap_port"
-        assert gap_port.renormalization_impedance == 50
-        gap_port.is_circuit_port = True
-        assert gap_port.is_circuit_port
-        edb.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_edb_statistics(self):
-        """Get statistics."""
-        edb = self.edb_examples.get_si_verse_sfp()
-        edb_stats = edb.get_statistics(compute_area=False)
-        assert edb_stats
-        assert edb_stats.num_layers
-        assert edb_stats.stackup_thickness
-        assert edb_stats.num_vias
-        # assert edb_stats.occupying_ratio
-        # assert edb_stats.occupying_surface
-        assert edb_stats.layout_size
-        assert edb_stats.num_polygons
-        assert edb_stats.num_traces
-        assert edb_stats.num_nets
-        assert edb_stats.num_discrete_components
-        assert edb_stats.num_inductors
-        assert edb_stats.num_capacitors
-        assert edb_stats.num_resistors
-
-        # assert edb_stats.occupying_ratio["1_Top"] == 0.301682
-        # assert edb_stats.occupying_ratio["Inner1(GND1)"] == 0.937467
-        # assert edb_stats.occupying_ratio["16_Bottom"] == 0.204925
-        edb.close(terminate_rpc_session=False)
-
-    def test_create_rlc_component(self):
-        """Create rlc components from pin"""
-        edb = self.edb_examples.get_si_verse()
-        pins = edb.components.get_pin_from_component("U1", "1V0")
-        ref_pins = edb.components.get_pin_from_component("U1", "GND")
-        assert edb.components.create([pins[0], ref_pins[0]], "test_0rlc", r_value=1.67, l_value=1e-13, c_value=1e-11)
-        assert edb.components.create([pins[0], ref_pins[0]], "test_1rlc", r_value=None, l_value=1e-13, c_value=1e-11)
-        assert edb.components.create([pins[0], ref_pins[0]], "test_2rlc", r_value=None, c_value=1e-13)
-        edb.close(terminate_rpc_session=False)
-
-    def test_create_rlc_boundary_on_pins(self):
-        """Create hfss rlc boundary on pins."""
-        edb = self.edb_examples.get_si_verse()
-        pins = edb.components.get_pin_from_component("U1", "1V0")
-        ref_pins = edb.components.get_pin_from_component("U1", "GND")
-        assert edb.excitation_manager.create_rlc_boundary_on_pins(
-            pins[0], ref_pins[0], rvalue=1.05, lvalue=1.05e-12, cvalue=1.78e-13
-        )
-        edb.close(terminate_rpc_session=False)
-
     def test_configure_hfss_analysis_setup_enforce_causality(self):
-        """Configure HFSS analysis setup."""
+        """Configure HFSS analysis setup enforce causality."""
         edb = self.edb_examples.get_si_verse()
         assert len(edb.setups) == 0
         edb.hfss.add_setup()
@@ -436,186 +87,6 @@ class TestClass(BaseTestClass):
             sweep.enforce_causality = True
         setup.sweep_data = sweeps
         assert setup.sweep_data[0].enforce_causality
-        edb.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_create_various_ports_0(self):
-        """Create various ports."""
-        target_path = self.edb_examples.copy_test_files_into_local_folder("edb_edge_ports.aedb")[0]
-        edb = self.edb_examples.load_edb(target_path)
-        if edb.grpc:
-            prim_1_id = [i.edb_uid for i in edb.layout.primitives if i.net.name == "trace_2"][0]
-            assert edb.excitation_manager.create_edge_port_vertical(prim_1_id, ["-66mm", "-4mm"], "port_ver")
-            #  test issue #2151 fix -> control that edge port is assignined at desired position when single point passed
-            edge = edb.ports["port_ver"].core.edges[0]
-            assert edge.point == (-0.066, -0.004)
-        else:
-            # This method is also available at same location in grpc but is deprecated.
-            prim_1_id = [i.id for i in edb.modeler.primitives if i.net.name == "trace_2"][0]
-            assert edb.excitation_manager.create_edge_port_vertical(prim_1_id, ["-66mm", "-4mm"], "port_ver")
-
-        prim_2_id = [i.id for i in edb.layout.primitives if i.net.name == "trace_3"][0]
-        if edb.grpc:
-            assert edb.excitation_manager.create_edge_port_horizontal(
-                prim_1_id, ["-60mm", "-4mm"], prim_2_id, ["-59mm", "-4mm"], "port_hori", 30, "Lower"
-            )
-        else:
-            # This method is also available at same location in grpc but is deprecated.
-            assert edb.excitation_manager.create_edge_port_horizontal(
-                prim_1_id, ["-60mm", "-4mm"], prim_2_id, ["-59mm", "-4mm"], "port_hori", 30, "Lower"
-            )
-        if edb.grpc:
-            assert edb.excitation_manager.get_ports_number() == 2
-        else:
-            assert edb.hfss.get_ports_number() == 2
-        port_ver = edb.ports["port_ver"]
-        assert not port_ver.is_null
-        assert not port_ver.is_circuit_port
-        assert port_ver.boundary_type in ["PortBoundary", "port"]  # grpc returns 'port'
-
-        port_hori = edb.ports["port_hori"]
-        assert port_hori.reference_terminal
-
-        kwargs = {
-            "layer_name": "Top",
-            "net_name": "SIGP",
-            "width": "0.1mm",
-            "start_cap_style": "Flat",
-            "end_cap_style": "Flat",
-        }
-        traces = []
-        trace_paths = [
-            [["-40mm", "-10mm"], ["-30mm", "-10mm"]],
-            [["-40mm", "-10.2mm"], ["-30mm", "-10.2mm"]],
-            [["-40mm", "-10.4mm"], ["-30mm", "-10.4mm"]],
-        ]
-        for p in trace_paths:
-            t = edb.modeler.create_trace(path_list=p, **kwargs)
-            traces.append(t)
-
-        paths_ids = [i.id for i in traces]
-        pts = [i.center_line[0] for i in traces]
-        if edb.grpc:
-            wave_port = edb.excitation_manager.create_bundle_wave_port(paths_ids, pts)
-        else:
-            wave_port = edb.excitation_manager.create_bundle_wave_port(paths_ids, pts)
-        wave_port.horizontal_extent_factor = 10
-        assert wave_port.horizontal_extent_factor == 10
-        wave_port.vertical_extent_factor = 10
-        wave_port.radial_extent_factor = 1
-        assert wave_port.radial_extent_factor == 1
-        wave_port.pec_launch_width = "0.02mm"
-        assert wave_port.pec_launch_width
-        assert not wave_port.deembed
-        assert wave_port.deembed_length == 0.0
-        # TODO check bug pyedb-core #675
-        # wave_port.do_renormalize = True
-        # assert wave_port.do_renormalize
-        # wave_port.do_renormalize = False
-        assert not wave_port.do_renormalize
-        if edb.grpc:
-            assert edb.excitation_manager.create_differential_wave_port(
-                traces[1].id,
-                trace_paths[0][0],
-                traces[2].id,
-                trace_paths[1][0],
-                horizontal_extent_factor=8,
-                port_name="df_port",
-            )
-        else:
-            assert edb.excitation_manager.create_differential_wave_port(
-                traces[1].id,
-                trace_paths[0][0],
-                traces[2].id,
-                trace_paths[1][0],
-                horizontal_extent_factor=8,
-                port_name="df_port",
-            )
-        assert not edb.ports["df_port"].is_null
-        p, n = edb.ports["df_port"].terminals
-        assert p.name == "df_port:T1"
-        assert n.name == "df_port:T2"
-        edb.ports["df_port"].decouple()
-        p.couple_ports(n)
-
-        traces_id = [i.id for i in traces]
-        paths = [i[1] for i in trace_paths]
-        if edb.grpc:
-            df_port = edb.excitation_manager.create_bundle_wave_port(traces_id, paths)
-        else:
-            df_port = edb.excitation_manager.create_bundle_wave_port(traces_id, paths)
-        assert df_port.name
-        assert df_port.terminals
-        df_port.horizontal_extent_factor = 10
-        df_port.vertical_extent_factor = 10
-        df_port.deembed = True
-        df_port.deembed_length = "1mm"
-        assert df_port.horizontal_extent_factor == 10
-        assert df_port.vertical_extent_factor == 10
-        assert df_port.deembed
-        assert df_port.deembed_length == 1e-3
-        edb.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_create_various_ports_1(self):
-        """Create various ports."""
-        target_path = self.edb_examples.copy_test_files_into_local_folder("edb_edge_ports.aedb")[0]
-        edb = self.edb_examples.load_edb(target_path)
-        kwargs = {
-            "layer_name": "TOP",
-            "net_name": "SIGP",
-            "width": "0.1mm",
-            "start_cap_style": "Flat",
-            "end_cap_style": "Flat",
-        }
-        traces = []
-        trace_paths = [
-            [["-40mm", "-10mm"], ["-30mm", "-10mm"]],
-            [["-40mm", "-10.2mm"], ["-30mm", "-10.2mm"]],
-            [["-40mm", "-10.4mm"], ["-30mm", "-10.4mm"]],
-        ]
-        for p in trace_paths:
-            t = edb.modeler.create_trace(path_list=p, **kwargs)
-            traces.append(t)
-
-        if config["use_grpc"]:
-            assert edb.excitation_manager.create_wave_port(traces[0].id, trace_paths[0][1], "wave_port")
-        else:
-            assert edb.excitation_manager.create_wave_port(traces[0], trace_paths[0][0], "wave_port")
-
-        assert edb.excitation_manager.create_differential_wave_port(
-            traces[0],
-            trace_paths[0][0],
-            traces[1],
-            trace_paths[1][0],
-            horizontal_extent_factor=8,
-        )
-
-        paths = [i[1] for i in trace_paths]
-        if config["use_grpc"]:
-            p = edb.excitation_manager.create_bundle_wave_port(traces, paths, port_name="port2")
-        else:
-            p = edb.excitation_manager.create_bundle_wave_port(traces, paths)
-        p.horizontal_extent_factor = 6
-        p.vertical_extent_factor = 5
-        p.pec_launch_width = "0.02mm"
-        p.radial_extent_factor = 1
-        assert p.horizontal_extent_factor == 6
-        assert p.vertical_extent_factor == 5
-        assert p.pec_launch_width == "0.02mm"
-        assert p.radial_extent_factor == 1
-        edb.close(terminate_rpc_session=False)
-
-    def test_set_all_antipad_values(self):
-        """Set all anti-pads from all pad-stack definition to the given value."""
-        edb = self.edb_examples.get_si_verse()
-        assert edb.padstacks.set_all_antipad_value(0.0)
         edb.close(terminate_rpc_session=False)
 
     @pytest.mark.skipif(not config["use_grpc"] and is_linux, reason="Crash on DotNet and gRPC on Linux.")
@@ -709,7 +180,7 @@ class TestClass(BaseTestClass):
 
     @pytest.mark.skipif(not config["use_grpc"] and is_linux, reason="Crash on DotNet and gRPC on Linux.")
     def test_hfss_simulation_setups_consolidation(self):
-        """Create a setup from a template and evaluate its properties."""
+        """Create a setup from a template and evaluate settings consolidation."""
         edbapp = self.edb_examples.get_si_verse()
         setup1 = edbapp.hfss.add_setup("setup1")
         assert setup1.set_solution_single_frequency()
@@ -796,6 +267,7 @@ class TestClass(BaseTestClass):
         edbapp.close(terminate_rpc_session=False)
 
     def test_siwaves_simulation_setups_consolidation(self):
+        """Create a SIWave setup and evaluate all settings consolidation."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_siwave_setup()
         setup = edbapp.setups[setup.name]
@@ -819,8 +291,6 @@ class TestClass(BaseTestClass):
         assert not adv_settings.include_split_plane_coupling
         adv_settings.inf_gnd_location = 1e-3
         assert adv_settings.inf_gnd_location == 1e-3
-        # TODO check why this line is crashing
-        # adv_settings.max_coupled_lines = 30
         adv_settings.mesh_automatic = False
         assert not adv_settings.mesh_automatic
         adv_settings.mesh_frequency = 30e9
@@ -903,16 +373,16 @@ class TestClass(BaseTestClass):
         assert not sp.use_state_space
 
         assert setup.name == "test_siwave_setup"
-
         edbapp.close(terminate_rpc_session=False)
 
     def test_siwaves_dcir_simulation_setups_consolidation(self):
+        """Create a SIWave DCIR setup and evaluate all settings."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_siwave_dcir_setup()
         setup.name = "test_siwave_dcir_setup"
         assert setup.name == "test_siwave_dcir_setup"
 
-        # DC settings (includes DCIR-specific properties via SIWaveDCIRDCSettings)
+        # DC settings
         dc = setup.settings.dc
         dc.compute_inductance = True
         assert dc.compute_inductance
@@ -945,7 +415,7 @@ class TestClass(BaseTestClass):
         dc.via_report_path = "via_report.txt"
         assert dc.via_report_path == "via_report.txt"
 
-        # dc advanced
+        # DC advanced
         dc_adv = setup.settings.dc_advanced
         dc_adv.dc_min_plane_area_to_mesh = "0.30mm2"
         assert dc_adv.dc_min_plane_area_to_mesh == "0.30mm2"
@@ -974,7 +444,7 @@ class TestClass(BaseTestClass):
         dc_adv.refine_vias = True
         assert dc_adv.refine_vias
 
-        # general (for DCIR setups, general proxies to DC settings for backward compat)
+        # General (for DCIR setups, proxies to DC settings for backward compat)
         general = setup.settings.general
         general.compute_inductance = True
         assert general.compute_inductance
@@ -989,12 +459,14 @@ class TestClass(BaseTestClass):
         edbapp.close(terminate_rpc_session=False)
 
     def test_raptor_x_simulation_setups_consolidation(self):
+        """Create a RaptorX setup and evaluate all settings."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_raptor_x_setup(name="test_raptorx_setup")
         setup.name = "test_raptorx_setup"
         assert setup.name == "test_raptorx_setup"
-        assert not setup.sweep_data  # default we don't create sweep if not data provided while setup creation
-        # advanced general settings
+        assert not setup.sweep_data
+
+        # Advanced general settings
         adv_settings = setup.settings.advanced
         adv_settings.auto_removal_sliver_poly = 1e-2
         assert adv_settings.auto_removal_sliver_poly == 1e-2
@@ -1011,11 +483,6 @@ class TestClass(BaseTestClass):
         if config["use_grpc"]:
             adv_settings.net_settings_options = {"VDD": ["mesh_vias", "via_diameter"]}
             assert adv_settings.net_settings_options == {"VDD": ["mesh_vias", "via_diameter"]}
-        else:
-            # Todo
-            pass
-            # adv_settings.net_settings_options = [["VDD", "mesh_vias"], ["VDD", "via_diameter"]]
-            # assert adv_settings.net_settings_options == {"VDD": ["mesh_vias", "via_diameter"]}
         adv_settings.override_shrink_factor = 2.0
         assert adv_settings.override_shrink_factor == 2.0
         adv_settings.plane_projection_factor = 2.0
@@ -1049,7 +516,7 @@ class TestClass(BaseTestClass):
         adv_settings.use_relaxed_z_axis = True
         assert adv_settings.use_relaxed_z_axis
 
-        # general settings
+        # General settings
         general = setup.settings.general
         general.global_temperature = 30.0
         assert general.global_temperature == 30.0
@@ -1067,6 +534,7 @@ class TestClass(BaseTestClass):
 
     @pytest.mark.skipif(not config["use_grpc"], reason="issue #1860. Missing create_q3d_setup in dotnet")
     def test_q3d_simulation_setups_consolidation(self):
+        """Create a Q3D setup and evaluate all settings."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_q3d_setup()
         assert not setup.is_null
@@ -1074,7 +542,7 @@ class TestClass(BaseTestClass):
         assert setup.name == "test_q3d_setup"
         assert setup.setup_type == "q3d"
 
-        # acrl settings
+        # ACRL settings
         acrl = setup.settings.acrl
         acrl.max_frequency = 20
         assert acrl.max_frequency == 20
@@ -1087,7 +555,7 @@ class TestClass(BaseTestClass):
         acrl.percent_error = 0.5
         assert acrl.percent_error == 0.5
 
-        # advanced meshing settings
+        # Advanced meshing settings
         adv_mesh = setup.settings.advanced_meshing
         adv_mesh.arc_step_size = 0.2
         assert adv_mesh.arc_step_size == 0.2
@@ -1095,19 +563,13 @@ class TestClass(BaseTestClass):
         assert adv_mesh.arc_to_chord_error == 1e-6
         adv_mesh.circle_start_azimuth = 15
         assert adv_mesh.circle_start_azimuth == 15
-        # TODO check pyedb-core bug #682 -> setter is broken
-        # adv_mesh.layer_alignment = 1e-4
-        # assert adv_mesh.layer_alignment == 1e-4
         adv_mesh.max_num_arc_points = 6
         assert adv_mesh.max_num_arc_points == 6
         adv_mesh.use_arc_chord_error_approx = True
         assert adv_mesh.use_arc_chord_error_approx
 
-        # cg settings
+        # CG settings
         cg = setup.settings.cg
-        # TODO check pyedb-core bug #683 -> setter is broken
-        # cg.auto_incr_sol_order = False
-        # assert not cg.auto_incr_sol_order
         cg.compression_tol = 1e-6
         assert cg.compression_tol == 1e-6
         cg.max_passes = 20
@@ -1123,7 +585,7 @@ class TestClass(BaseTestClass):
         cg.solution_order = "higher"
         assert cg.solution_order == "higher"
 
-        # dcrl settings
+        # DCRL settings
         dcrl = setup.settings.dcrl
         dcrl.max_passes = 20
         assert dcrl.max_passes == 20
@@ -1138,7 +600,7 @@ class TestClass(BaseTestClass):
         dcrl.solution_order = "higher"
         assert dcrl.solution_order == "higher"
 
-        # general settings
+        # General settings
         general = setup.settings.general
         general.do_ac = True
         assert general.do_ac
@@ -1155,6 +617,7 @@ class TestClass(BaseTestClass):
         edbapp.close(terminate_rpc_session=False)
 
     def test_sweep(self):
+        """Create a sweep and evaluate its properties."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_hfss_setup(
             name="test_hfss_setup", distribution="log_scale", start_freq="0GHz", stop_freq="1GHz", freq_step="10"
@@ -1185,6 +648,7 @@ class TestClass(BaseTestClass):
 
     @pytest.mark.skipif(is_linux and not config["use_grpc"], reason="Randomly fails on linux dotnet")
     def test_siwave_simulation_setup_dotnet_compatibility(self):
+        """Validate SIWave DCIR setup settings attribute compatibility."""
         edbapp = self.edb_examples.get_si_verse()
         setup = edbapp.simulation_setups.create_siwave_dcir_setup("setup_1")
 
@@ -1291,7 +755,6 @@ class TestClass(BaseTestClass):
             setup_2 = edbapp.simulation_setups.setups["setup_1"]
         else:
             setup_2 = setup
-
         assert setup_2.settings.dc_advanced.dc_min_plane_area_to_mesh == "0.30mm2"
         assert dc_adv.dc_min_void_area_to_mesh == "0.02mm2"
         assert dc_adv.energy_error == 1.5
@@ -1306,27 +769,21 @@ class TestClass(BaseTestClass):
         assert dc_adv.refine_bondwires
         assert dc_adv.refine_vias
 
-        # General settings (backward compat — proxies to DC settings for DCIR setups)
+        # General settings (backward compat)
         general = settings.general
         general.pi_slider_position = 0
         general.si_slider_position = 0
         general.use_custom_settings = False
         general.use_si_settings = False
 
-        # -------------------------
-        # Validate settings (ASSERTS)
-        # -------------------------
         if config["use_grpc"]:
             setup_2 = edbapp.simulation_setups.setups["setup_1"]
         else:
             setup_2 = setup
-        # General assertions (backward compat)
         assert setup_2.settings.general.pi_slider_position == 0
         assert setup_2.settings.general.si_slider_position == 0
         assert not setup_2.settings.general.use_custom_settings
         assert not setup_2.settings.general.use_si_settings
-
-        # test syz setup
 
         if config["use_grpc"]:
             setup_2 = edbapp.simulation_setups.setups["setup_1"]
@@ -1335,88 +792,14 @@ class TestClass(BaseTestClass):
         assert "pi_slider_position", "si_slider_position" in setup_2.get_configurations().items()
         edbapp.close(terminate_rpc_session=False)
 
-    @pytest.mark.skipif(
-        config["use_grpc"] and ansys.edb.core.__version__ == "0.2.6",
-        reason="Test skipped for ansys-edb-core version 0.2.6",
-    )
-    def test_edb_settings(self):
-        edbapp = self.edb_examples.get_si_verse()
-        assert type(edbapp.logger) == EdbLogger
-        assert edbapp.version == config["desktopVersion"]
-        assert edbapp.base_path == edbapp.ansys_em_path
-        assert edbapp.grpc == config["use_grpc"]
-        val = edbapp.value(1)
-        assert val
-        edbapp.add_design_variable("test", 1)
-        assert edbapp.design_variables["test"].value == 1
-        edbapp.add_project_variable("test2", 2)
-        assert edbapp.project_variables["$test2"].value == 2
-        assert type(edbapp.layout_validation)
-        # TODO add layout validation obj extended test
-        assert "test", "$test2" in edbapp.variables.items()
-        edbapp.change_design_variable_value("test", 3)
-        assert edbapp.design_variables["test"].value == 3
-        assert edbapp.get_bounding_box()
-        assert edbapp.get_statistics()
-        assert edbapp.are_port_reference_terminals_connected()
-        edbapp.close(terminate_rpc_session=False)
-
-    @pytest.mark.skipif(
-        config["use_grpc"] and config["desktopVersion"] < "2026.1", reason="working with latest release"
-    )
-    def test_ports_and_sources_creation(self):
-        edbapp = self.edb_examples.get_si_verse()
-        p1 = edbapp.padstacks.instances_by_name["Via1"].create_terminal("p1")
-        p2 = edbapp.padstacks.instances_by_name["Via2"].create_terminal("p2")
-        edbapp.create_port(p1, p2, True, "test")
-        assert edbapp.ports["test"]
-        p3 = edbapp.padstacks.instances_by_name["Via3"].create_terminal("p3")
-        p4 = edbapp.padstacks.instances_by_name["Via4"].create_terminal("p4")
-        edbapp.create_port(p3, p4, False, "test2")
-        assert edbapp.ports["test2"]
-
     @pytest.mark.skipif(not config["use_grpc"], reason="DotNet skipping")
     def test_siwave_simulation_setup_bug(self):
+        """Validate SIWave DCIR setup use_loop_res_for_per_pin persists after reload."""
         edbapp = self.edb_examples.create_empty_edb()
         setup = edbapp.simulation_setups.create_siwave_dcir_setup("setup_1")
         settings = setup.settings
         settings.use_loop_res_for_per_pin = False
 
         setup = edbapp.setups["setup_1"]
-        assert not setup.settings.use_loop_res_for_per_pin  # fail on .net
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_horizontal_wave_ports(self):
-        local_path = Path(__file__).parent.parent
-        example_folder = os.path.join(local_path, "example_models", "TEDB")
-        source_path_edb = os.path.join(example_folder, "example_arbitrary_wave_ports.aedb")
-        edbapp = self.edb_examples.load_edb(source_path_edb)
-        voids = edbapp.layout.primitives[0].voids
-        for void in voids:
-            edbapp.excitation_manager.create_horizontal_wave_port(void)
-        assert len(edbapp.ports) == 6
-        edbapp.close(terminate_rpc_session=False)
-
-    def test_load_multiple_edb(self):
-        edb1 = self.edb_examples.get_si_board()
-        assert len(list(edb1.components.instances.values())) == 4
-        assert list(edb1.components.instances.values())[0].name == "U0"
-        edb2 = self.edb_examples.get_si_verse()
-        assert list(edb1.components.instances.values())[0].name == "U0"
-        assert len(list(edb2.components.instances.values())) == 509
-        assert list(edb2.components.instances.values())[0].name == "C380"
-        edb1.close(terminate_rpc_session=False)  # testing server RPC should not be closed
-        assert edb1.active_cell is None
-        assert len(list(edb2.components.instances.values())) == 509
-        assert list(edb2.components.instances.values())[0].name == "C380"
-        edb2.close(terminate_rpc_session=False)
-        assert edb2.active_cell is None
-
-    def test_etching_on_nets(self):
-        edbapp = self.edb_examples.get_si_verse()
-        for layer in list(edbapp.stackup.signal_layers.values()):
-            layer.etch_factor = 0.1
-            layer.etch_net_class = "no_power_ground"
-            assert layer.etch_net_class == "no_power_ground"
-
+        assert not setup.settings.use_loop_res_for_per_pin
         edbapp.close(terminate_rpc_session=False)
