+++ tests/system/test_edb_stackup.py
@@ -1298,3 +1298,12 @@ class TestClass(BaseTestClass):
         outline_layer = non_stackup_layers["Outline"]
         assert not outline_layer.is_stackup_layer
         edbapp.close(terminate_rpc_session=False)
+
+    def test_etching_on_nets(self):
+        """Evaluate etch factor and etch net class on signal layers."""
+        edbapp = self.edb_examples.get_si_verse()
+        for layer in list(edbapp.stackup.signal_layers.values()):
+            layer.etch_factor = 0.1
+            layer.etch_net_class = "no_power_ground"
+            assert layer.etch_net_class == "no_power_ground"
+        edbapp.close(terminate_rpc_session=False)
